backend/requirements.txt
fastapi
uvicorn
pydantic[email]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
scikit-learn
pandas
numpy
shap
lime
torch
transformers
python-dotenv
httpx
